requirements.txt
fyers-apiv3
python-dotenv
pandas
requests
langgraph
langchain-core
matplotlib
